fix(cli): address detailed scan review blockers
